refactor(common): unify `CodePoint` representation to `Int` and improve multiplatform consistency
- Replaced `UInt` with `Int` for `CodePoint` to streamline type handling across platforms.
- Consolidated `Util` functions for `computeValue` and `computeValueIfAbsent` into `nonJvm` module and removed redundant platform-specific variants.
- Simplified logic in BMP and supplementary character handling in `stringValue`, `charCount`, and `category` methods.
- Adjusted tests in `CodePointTest` to reflect the changes and ensure behavior remains consistent across edge cases and platforms.

+++ codegentle-common/src/jsBasedMain/kotlin/love/forte/codegentle/common/codepoint/CodePoint.jsBased.kt
@@ -8,9 +8,9 @@ internal expect fun jsFromCodePoint(code: Int): String
 
 @InternalCodePointApi
 public actual fun String.codePointAt(index: Int): CodePoint {
-    val code = jsCodePointAt(this, index).toUInt()
+    val code = jsCodePointAt(this, index)
     return CodePoint(code)
 }
 
 internal actual fun CodePoint.stringValue(): String =
-    jsFromCodePoint(code.toInt())
+    jsFromCodePoint(code)
